move non layout components to shared folder
