+++ imports/blocks/give/fieldsets/confirm/PaymentOptionsLayout.js
@@ -1,90 +1,11 @@
 import { PropTypes } from "react";
-import AccountType from "../../../../components/accountType";
+import { SavedAccount } from "../shared";
 
 const Header = () => (
   <h4 className="text-center flush-bottom">
     Change Payment Account
   </h4>
 );
-
-const AccountNumber = ({ accountNumber }) => (
-  <span>
-    {accountNumber
-      .slice(0, accountNumber.length - 5)
-      .replace(/./gmi, "*")
-    }
-    {accountNumber.slice(-4)}
-  </span>
-);
-
-AccountNumber.propTypes = {
-  accountNumber: PropTypes.string,
-};
-
-const SavedAccount = ({
-  account,
-  choose,
-  savedAccount,
-}) => (
-  <div
-    style={{ position: "relative", cursor: "pointer" }}
-    id={account.id}
-    onClick={choose}
-  >
-    <div
-      className={
-        "soft-ends push-double-left text-left hard-right " +
-        "outlined--light outlined--bottom relative"
-      }
-    >
-      <div className="display-inline-block soft-half-ends one-whole">
-        <h6 className="flush-bottom float-left text-dark-tertiary">
-          {account.name}
-        </h6>
-      </div>
-
-      <h5 className="hard one-whole flush-bottom text-dark-tertiary">
-        <AccountNumber accountNumber={account.payment.accountNumber} />
-        <span className="float-right soft-half-left">
-          <AccountType
-            width="40px"
-            height="25px"
-            type={account.payment.paymentType}
-          />
-        </span>
-      </h5>
-    </div>
-    <div className="locked-ends locked-sides">
-      <input
-        type="checkbox"
-        id={`label${account.id}`}
-        readOnly
-        checked={Number(account.id) === Number(savedAccount.id)}
-        style={{
-          opacity: 0,
-          position: "absolute",
-          top: 0,
-          left: 0,
-          padding: "50px",
-        }}
-      />
-      <label
-        htmlFor={`label${account.id}`}
-        style={{
-          position: "absolute",
-          top: "50%",
-          left: 0,
-        }}
-      />
-    </div>
-  </div>
-);
-
-SavedAccount.propTypes = {
-  account: PropTypes.object,
-  choose: PropTypes.func,
-  savedAccount: PropTypes.object,
-};
 
 const PaymentOptionsLayout = ({
   changeAccounts,
