+++ imports/blocks/give/fieldsets/confirm/TransactionLayout.js
@@ -1,9 +1,7 @@
 import { PropTypes } from "react";
 import {
-  ButtonText,
-  cardType,
-  Icon,
-  isIOS,
+  ActionButton,
+  ListItem,
   monetize,
   PaymentOptions,
 } from "../shared";
@@ -20,86 +18,6 @@ const Header = ({ override, personal }) => {
 Header.propTypes = {
   override: PropTypes.object,
   personal: PropTypes.object,
-};
-
-const ListItem = ({
-  transaction,
-}) => (
-  <div className="soft-half-ends hard-sides">
-
-    <div className="grid" style={{ verticalAlign: "middle" }}>
-
-      <div className="grid__item two-thirds" style={{ verticalAlign: "middle" }}>
-        <h5 className="text-dark-secondary flush text-left">
-          {transaction.label}
-        </h5>
-      </div>
-
-      <div className="grid__item one-third text-right" style={{ verticalAlign: "middle" }}>
-        <h5 className="text-dark-secondary flush">
-          {monetize(transaction.value)}
-        </h5>
-      </div>
-
-    </div>
-  </div>
-);
-
-ListItem.propTypes = {
-  transaction: PropTypes.object,
-};
-
-const ActionButton = ({
-  completeGift,
-  payment,
-  savedAccount,
-  schedules,
-  scheduleToRecover,
-}) => {
-  if (isIOS()) {
-    return (
-      <div>
-        <p className="text-dark-secondary">
-          <small>
-            <em>
-              Due to restrictions with your operating system,
-              you must complete your gift in the browser.
-            </em>
-          </small>
-        </p>
-        <button
-          className="btn soft-half-top one-whole"
-          onClick={completeGift}
-        >
-          Complete Gift in Browser
-        </button>
-      </div>
-    );
-  }
-  return (
-    <button className="btn soft-half-top one-whole" type="submit">
-      <ButtonText
-        payment={payment}
-        savedAccount={savedAccount}
-        schedules={schedules}
-        scheduleToRecover={scheduleToRecover}
-      />
-      &nbsp;
-      <Icon
-        cardType={
-          cardType(payment, savedAccount)
-        }
-      />
-    </button>
-  );
-};
-
-ActionButton.propTypes = {
-  completeGift: PropTypes.func,
-  payment: PropTypes.object,
-  savedAccount: PropTypes.object,
-  schedules: PropTypes.array,
-  scheduleToRecover: PropTypes.bool,
 };
 
 const TransactionLayout = ({
