+++ imports/blocks/give/fieldsets/payment/Layout.js
@@ -1,12 +1,12 @@
 import { PropTypes } from "react";
 
 import Controls from "../../../../components/controls";
-import Forms from "../../../../components/forms";
-import { creditCard } from "../../../../util/format";
 
 import {
-  cardType,
-  Icon,
+  BankFields,
+  CardFields,
+  SavePaymentCheckBox,
+  SavePaymentInput,
 } from "../shared";
 
 const DEFAULT_TOGGLES = ["Credit Card", "Bank Account"];
@@ -24,224 +24,6 @@ const Header = ({
 
 Header.propTypes = {
   override: PropTypes.object,
-};
-
-const BankFields = ({
-  payment,
-  saveData,
-  validate,
-}) => {
-  if (payment.type !== "ach") return null;
-  return (
-    <div>
-      <Forms.Input
-        id="routingNumber"
-        name="billing-routing-number"
-        label="Routing Number"
-        type="tel"
-        errorText="Please enter your routing number"
-        defaultValue={payment.routingNumber}
-        onChange={saveData}
-        validation={validate}
-        autoFocus
-      />
-
-      <Forms.Input
-        id="accountNumber"
-        name="billing-account-number"
-        label="Account Number"
-        type="tel"
-        errorText="Please enter your account number"
-        defaultValue={payment.accountNumber}
-        onChange={saveData}
-        validation={validate}
-      />
-
-
-      <div className="grid">
-
-        <div className="grid__item one-whole">
-          <Forms.Select
-            name="billing-account-type"
-            id="accountType"
-            label="Account Type"
-            onChange={saveData}
-            validation={validate}
-            defaultValue={payment.accountType}
-            errorText="Please choose your account type"
-            includeBlank
-            items={[
-              { value: "checking", label: "Checking" },
-              { value: "savings", label: "Savings" },
-            ]}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-BankFields.propTypes = {
-  payment: PropTypes.object,
-  saveData: PropTypes.func,
-  validate: PropTypes.func,
-};
-
-const CardFields = ({
-  formatExp,
-  payment,
-  saveData,
-  savedAccount,
-  validate,
-}) => {
-  if (payment.type === "ach") return null;
-  return (
-    <div>
-      <Forms.Input
-        name="billing-cc-number"
-        id="cardNumber"
-        label="Card Number"
-        type="tel"
-        errorText="Please enter your card number"
-        defaultValue={payment.cardNumber}
-        format={creditCard}
-        onChange={saveData}
-        validation={validate}
-      >
-        <div className="locked locked-right soft-double-right locked-top" style={{ top: "-3px" }}>
-          <RenderIcon
-            payment={payment}
-            savedAccount={savedAccount}
-          />
-        </div>
-      </Forms.Input>
-      <div className="grid">
-        <div className="grid__item one-half">
-          <Forms.Input
-            id="expiration"
-            name="billing-cc-exp"
-            label="Exp (MM/YY)"
-            type="tel"
-            errorText="Please enter a valid expiration number"
-            defaultValue={payment.expiration}
-            format={formatExp}
-            onChange={saveData}
-            validation={(value) => (value.length > 0)}
-            data-expiry-input
-          />
-        </div>
-        <div className="grid__item one-half">
-          <Forms.Input
-            id="ccv"
-            name="billing-cvv"
-            label="CCV"
-            type="tel"
-            errorText="Please enter a valid ccv number"
-            defaultValue={payment.ccv}
-            onChange={saveData}
-            validation={validate}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-CardFields.propTypes = {
-  formatExp: PropTypes.func,
-  payment: PropTypes.object,
-  saveData: PropTypes.func,
-  savedAccount: PropTypes.object,
-  validate: PropTypes.func,
-};
-
-const RenderIcon = ({
-  payment,
-  savedAccount,
-}) => {
-  const masked = payment.type === "ach" ? payment.accountNumber : payment.cardNumber;
-  if (!masked) return null;
-
-  const paymentType = cardType(payment, savedAccount);
-  if (!paymentType) return null;
-
-  // replace with SVG
-  return <Icon cardType={paymentType} />;
-};
-
-RenderIcon.propTypes = {
-  payment: PropTypes.object,
-  savedAccount: PropTypes.object,
-};
-
-const SavePaymentCheckBox = ({
-  savedAccount,
-  savePayment,
-  shouldSaveState,
-  schedules,
-  transactionType,
-}) => {
-  if (
-    !savedAccount.id &&
-    transactionType !== "guest" &&
-    Object.keys(schedules).length === 0
-  ) {
-    return (
-      <Forms.Checkbox
-        name="savePayment"
-        defaultValue={shouldSaveState}
-        clicked={savePayment}
-      >
-        Save this payment for future contributions
-      </Forms.Checkbox>
-    );
-  }
-  return null;
-};
-
-SavePaymentCheckBox.propTypes = {
-  savedAccount: PropTypes.object,
-  savePayment: PropTypes.func,
-  shouldSaveState: PropTypes.bool,
-  schedules: PropTypes.array,
-  transactionType: PropTypes.string,
-};
-
-const SavePaymentInput = ({
-  saveName,
-  savedAccount,
-  schedules,
-  shouldSaveState,
-  payment,
-  transactionType,
-}) => {
-  if (
-    shouldSaveState &&
-    !savedAccount.id &&
-    transactionType !== "guest" &&
-    Object.keys(schedules).length === 0
-  ) {
-    return (
-      <Forms.Input
-        name="accountName"
-        label="Saved Account Name"
-        classes={["soft-bottom", "flush-bttom"]}
-        errorText="Please enter a name for the account"
-        validation={saveName}
-        defaultValue={payment.type === "ach" ? "Bank Account" : "Credit Card"}
-      />
-    );
-  }
-  return null;
-};
-
-SavePaymentInput.propTypes = {
-  saveName: PropTypes.func,
-  savedAccount: PropTypes.object,
-  schedules: PropTypes.array,
-  shouldSaveState: PropTypes.bool,
-  payment: PropTypes.object,
-  transactionType: PropTypes.string,
 };
 
 const NextButton = ({
